Move RawDataView to package to improve readability
- Introduces the raw_data_view package to help isolate strict concerns
  of RawDataView
- Reduces the RawDataView Python footprint so it is easier to read
