+++ src/le_beta_vis/frontend/views/raw_data_view/_RawDataViewStyle.py
